refactor: move artworksavebutton to its own component
